Removes kref_next
They are using these as a source of integer IDs, not as references.

Plan9 does this all over the place too.  Here, just an increment of an
int is fine.  In places like #s, we'll need an ID service (same idea
behind PIDs).  Solaris had 'vmem' for this, and it was pretty cool.