waserror() audit
Our waserrors in the error case need a poperror or nexterror.  This
isn't how plan9 works.

This was rather painful.  Enough so that perhaps we should have changed
our ways.  Maybe next time we port a plan9 derivative.

I also dealt with the weird errstr shit in namec the same way I did with
nxm: commenting it out.  I found that (back then) to make namec harder
to debug.  We'll see.