Open mode checks, don't use == for OREAD
There's a disconnect between omode and the access bits part of the VFS
open flags.  Don't simply check for equality of OREAD.

I didn't want to change omode higher in the stack (to do so, we can pass
it through the poorly named 'openmode').  I wanted to keep the vfs_flags
info around as long as possible, in case lower layers want to make use
of that info.

Btw, OREAD's value is 0, which means we can't check for the bit.
Fucking terrible.