Fix glibc to build with newer makes (XCC)
My make is 4.0, which isn't one of their hardcoded versions.  I get
complaints about sed and makeinfo's versions too.  The version detection
for sed is wrong, for one thing.  I haven't fixed those - not sure it's
necessary.

It was necessary to put the change in the main configure file, and not
only in the .in.  In fact, I'm not sure the .in is getting autoconfed
into the main configure all the time.  My autoconf is supposedly good,
but I was sometimes getting the dreaded BP_SYM error, which is when our
base_os change isn't in the main configure.

Since we're having a configure file anyway, I put the base_os business
in there too.  These are the only two diffs from our configure.in and
the default glibc one.

You'll need to fully rebuild glibc for this to take effect.