Removes the old networking stack (XCC)
In addition to the socket family of glibc and syscall numbers, I also
removed the syscalls for read/write over serial/ethernet, and the
ethernet audio device.  If we need any of this in the future, we can
get it from git.

You need to fully rebuild glibc for this, including removing the entire
glibc-2.14.1/ directory (but not the -ros one).