Fix up eipconvtest.c
This runs under linux now and it prints out IP address strings.
The only problem is ... we don't know the right value. The unit
test did not include them. Ah well, at least it *seems* right.

This file will get copied to user libraries and then it's time to
fix it up for the kernel.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>