rtl8139 port
We don't support some of the ether->opts yet, probably due to a lack of
archether().  If we had it, the detection would make a little more sense
(they alloc a ctlr for every NIC, and then later find out via the opts
if it's an 8139).

We also don't do some of the mmu stuff, though maybe we should.