Sets ENOENT when appropriate
Userspace will often require ENOENT be set after syscalls fail in a
certain way.  For instance, cp stats the destination, and if it gets a
generic error, the program aborts.  It ought to get ENOENT (or success).

We'll run into this in any place where userspace makes decisions based
on errno.