The 8169 and mii from gpxe now build
Question: outl takes an int, but should probably take a
uintptr, or not? On MMIO systems it really should be 64-bit
thing, but on standard x86 it's 32 bits, and on (sigh) an alpha
it would be 64 bits without regard to it being IO or MMIO.

Need to redo the pci usage.

And, need to make it a real ether driver. I'm going to put the
basic structure in there for that now, based on the 8139.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>