Fixes #s
Looks like we won't be able to hang reference counted objects off of
c->aux, such that those objects change as we gen new entries.  The final
culprit was devclone, which copies aux, but doesn't incref.  And later,
it may or may not call the device's close.

We can still put reference counted things on c->aux, but they need to
not change during walks.  Check out devpipe and devip for examples.