Fixes netifgen
This was a problem from back in plan9, which I fixed in our NxM port.

The root issue is that netif code can't tell the difference between a
QID with a netid of 0 and one with no netid at all.  I think I fixed it.
I left in my comments for the next time we need to debug this.