Adds chaninfo()
This is useful for debugging.  Most devices will use devchaninfo.  But
if you want to spit out extra info, take a look at ipchaninfo.

This info will show up now when we do a "procinfo pid #" from the kernel
monitor.  We could also use it from procfs in the future.