More debugging for VMs.
Remove some of the chattier bits.

Add a new file, kern/kfs/bin/POR, which is a representation of
the F (bios) segment with one instruction, HLT.

Fix an incorrect test in root which would not allow the kernel to
read files (the check for user addresses should be in the system call).

Still not quite there.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>