Run a VM.
This set of changes, along with the runvm script,
let you run a vm.

To tell it's running: run /runvm script and hit return
a LOT. At some point you'll see a message like
GO FOR IT!

If you look at the RIP that gets printed out, you'll see
it's 0000fffe, which means it ran all the xorl instructions
and hit the infinite loop.

I'll be removing all this printk stuff when things are further along.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>