Early file mmaps are pinned and populated
for memory, pinning is currently irrelevant, since the kernel never
revokes memory.  populating is just a performance tradeoff: soft-page
faults at the first access, vs wasting large blobs of RAM (like if Go
decides to mmap a huge contig region, but doesn't use it) and greater
initial mmap costs.  both of these might change in the future.

for files, we run into correctness issues with blocking in vcore context
(which we don't allow).  for code and data that are touched from vcore
context, we need them to be both pinned and populated.  pinning prevents
further removal, and population ensures we never take a hard page fault.

since there is a lot of libc magic at startup and no good way (yet) to
identify the file segments needed by vcore context, i just pin/populate
any file mmap that happens before uthread or application code takes
over.

the question is what to do about all of the other mmaps out there,
memory and file alike.  the current default *flags* are unpinned,
unpopulated. it's up to userspace to change this if they want.

for memory, that means that mallocs are not populated/pinned.  as
mentioned above, we don't revoke anonymous mmaps, so in essence the
default already is pinned, but unpopulated, even though the MAP_LOCKED
flag isn't set.  i'm okay with this, but we'll need to look at all sorts
of mallocs (like uthread structs, syscall structs, etc) if this ever
changes.

for files mmapped later, including dlopens, userspace will need to pin
and populate.  populating from blocking media is going to be pretty
expensive, so either way we want userspace to be in control.  but since
the existing norm is to use MAP_LOCKED, and MAP_POPULATE, it makes sense
to stick with those.