Track the guest physical address used for VMs
It's pretty trivial to just track the guest pa in the
page struct, and it greatly simplifies VM code.

The alternatives are all ugly, as many of the extant VMs
out there show. We own this kernel, might as well make it easy.

pg_private was likely overcommited so I went this route.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>