MM populate cleanup
The main thing is the use of a history counter, which allows us to
detect changes in the VMRs, whether that change is just some prot
settings, or the existence of VMRs, etc.  While we can't reinspect the
vmr memory across an unlock-and-block, the history tells us it is safe
to keep following our previous instructions (keep mmaping a range of
pages).