SYS_populate_va (XCC)
Syscall for populating a range of virtual addresses.  This will usually
be used by PF handlers on file-backed VMRs, but it can be done for any
sort of user-controlled read-ahead that is desired.

If people want more control, like an FD interface that doesn't bother
with PTEs, let me know.

You need to reinstall your kernel headers.