2LS op for handling reflected faults
The pthread one knows how to handle page faults on file-backed virtual
addresses.

I'm using the existing pthread syscall handling, so this sits in
pthread.c for now.  Most other 2LSs will do the same thing.  In the
future, we could have a uthread helper for it, but only if we redo some
of the other syscall stuff.  Due to the ev_q and how we track syscalls,
which is a 2LS decision, most of this will probably stay in the specific
2LS.