Various APIC debugging and IOAPIC IRQ routing
Still busted, but getting a bit closer.  apiconline() needs work, and needs to
be called for all cores.  IOAPIC routing is broken, and the interfaces to IRQ
code are sloppy.  We have a weird mix of plan9 code and Akaros code that do
different things with PCI and IRQ handling.