Get the vm code to build again.
Due to other changes the code had stopped building.
So get it to at least build. We also know the
source of the vm startup problems and will fix them as
soon as we have working hardware or a working qemu.

Also, enable more vmx options for qemu. Nope, still not
enough there yet. Qemu doesn't seem to do what it says it will
do.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>