Make ndblib not reopen and reread files all the time.
Sadly, stat(2) returns 0 for mtime still. So we've got some
nasty hacks in here to make sure we only read files once. We'll
need to come back to this or better yet rewrite this mess in Go.
It's the 21st century and it's silly not to cache everything, and
it's super silly to write code like this in C.

Result:
cs&
mount -a '#s/cs' /net
csquery /net/cs tcp!localhost!22

The first time, it is slow, as it is now. Subsequent
csquery commands, if it hits the cache, are quite fast,
basically intantaneous, even in qemu.
This should make go test much more palatable.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>