Re-add the bsd library
We need it for netperf. I'm not in the mood to try to change
an autoconfig-based chamber of horrors at this point in my life.
Far easier to make a compatibility shim, and that alone should tell
you all you need to know about what's wrong with open source software
nowadays.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>