82563/i350: Fixes off-by-one issue
We were seeing TX underruns, which made me look closer.  I don't think
the -1s would work with the NEXT_RING macro, which expects the full
length of the ring.  Perhaps 9port had different ring macros?  Or they
wanted to skip that last entry for some reason...

Either way, we don't get the underruns with the removal of the -1s.