oprofile: Allow user mode to read oprofile data via '#K'.
Gee, that was easy. Now if there is data to read, it will be read
in standard oprofile format. This means we need to now build the
oprofile tools for Akaros. Also, time to get poor mans ftrace
so it will patch code for catching oprofile data.

I don't necessarily LIKE oprofile format; it's not endian-clean
and architecture-clean the way plan 9 tends to do it. But it's a
sub-standard now and we might as well just run with it.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>