oprofile: minimally working
We don't have on/off control; that's next.

You can echo something > '#K/kpoprofile' and it will go into
the buffer. You can than cat that file. It's a good way to test.
It works.

Need to get the programs going, see if they work and, if so,
enable the rest of the bits. Probably will implement the real
command interface too.

But, sigh. Will probably have to implement all the hokey oprofile
single-value-for-file files, just for compatibility with the
tools. Oh, yuck.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>