Fixes vcore counters
I was a little tossed up about putting the accounting in map/unmap or
not.  It'd be correct (if we did a little more unmapping for SCPs), but
in the end I wanted to parallelize the work as much as possible.  Our
'for each' vcore loops are probably long enough as it is.