Proc vcore list only has available vcores
Previously, you could see up to vcore 8 on an 8 core machine, even
though you could never get that last vcore.  This commit keeps the list
of vcores in sync with which vcores are available.

Right now, the rules for max_vcores is that it is set at process
creation and never changed.  The ksched could offer a process less than
the number of CG cores in the system, via the max_vcores() function.