oprofile
The oprofile tools are pretty hardwired to a certain directory structure
that is hard to recreate, and not clearly needed.

Change the output of #K/kpoprofile to the following:
PC
bits 63-4, time in ns; 3-0, number PCs following
[an array of PCs]

I don't think we'll miss the 1 ns precision, because there's no way
we're that accurate. +- 16 ns is fine.

TRACEME now calls oprofile_add_backtrace, which adds the backtrace.

Sadly, we're not seeing any backtrace data, but our Go program finds this output
very easy to digest.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>