Fixes KMSG IPI handling bug
Left over from the x86 change to not have IRQs globally disabled when running
IRQ handlers.

Note that even though RKMs are processed during a routine point, the lock still
needs to be irqsave.  The lock is grabbed from IRQ context, when sending
messages such as the alarm handler or the console IRQs.