oprofile: change data format for #K/kpoprofile
It now has a header descriptor, a time in ns, and 1
or more EIP values.

first word
high 8 bits is ee, which is an invalid address on amd64.
next 8 bits is protocol version
next 16 bits is unused, MBZ. Later, we can make it a packet type.
next 16 bits is core id
next 8 bits is unused
next 8 bits is # words following, should be > 0.

second word is time in ns.

Third and following words are PCs, there must be at least one of them.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>