Add a dispatcher for the alarm event
Every event has only a single event handler registered for it in the
ev_handlers array.  The alarm dispatcher sets ev_handler[EV_ALARM] to a
dispatcher function, which dispatches alarm events to registered
handlers keyed by alarmid.

This component ws built mostly in preparation for the addition of a
per-vcore alarm service that will need this dispatching functionality.