Remove the need for a poke_pvcalarm function
We wanted a way to avoid having to add the poke function to every
vcore_entry() so that they could take care of initializing their own
pvcalarm upon the first time they come up, and then starting it on
subsequent times if it happened to be disabled.  To do this we now just
preemptively initialize max_vcore() number of alarms (one for each
vcore) and start them / stop them upon enable/disable of the service.

Removes alot of complexity for very little extra overhead -- especially
on machines with a small number of max_vcores().