Fixes alarm pcpu printing
The old alarm code (from last week) had the sem and the func in separate
memory; now they are a union.  We were printing the func as if it was a
sem.

This change is also more useful, since we can see what the function
handler is (e.g. __ksched_tick) for non-semaphore based alarms.