Fix up alarm and pvcalarm tests
It appears that qeum doesn't properly virtualize the tsc accross vcores,
so calling read_tsc() in a situation where our vcores are actually
multiplexed on top of a single underlying linux task results in
incorrect values. The assertions in these tests have been changed to
make sure that AT LEAST the amount of time we expect has passed in real
time, rather than relying on the time counted per vcore.