Fixes network stats printf strings
All of these protocols are using u32s for their stats and other stuff.  The
format strings were looking for longs instead of ints.  This fixes them to use ints.

Alternatively, we could change the stats fields to u64.  I'm indifferent.

I didn't both changing any of the protocols that we don't use or even compile.