Only linearizeblock on #M qremoves
So we don't linearize blocks any time someone calls qremove.  The other
callers of qremove look okay.  It's up to the callers to know what to do
with the blocks.  For instance, bl2mem() knows how to deal with
extra_data.

Also, the other mount block op right above the change is now safer,
since pullupqueue() makes sure the full header is in place.